Split Address module out of Cardano.Api.Typed
