+++ cardano-api/cardano-api.cabal
@@ -36,6 +36,7 @@ library
                         Cardano.Api.TxSubmit.Types
 
                         -- Splitting up the big Typed module:
+                        Cardano.Api.Address
                         Cardano.Api.Eras
                         Cardano.Api.Error
                         Cardano.Api.Hash
